Modified and deleted some  __init__.py files
git_repo should be in the PYTHONPATH and is not intended to be a
python package.
git_repo/apps is for executables and scripts and should also not be a
package.
git_repo/kid_readout is a python package; I modified its __init__.py
to be blank since it was importing older code.