Refactored MeasurementList and added IOList.
MeasurementList is now the only container for sequences of Measurements, and it must be used to construct sequences that are passed as arguments to other Measurements.
Added IOList class, which has the same interface as a MeasurementList but is linked to a saved instance on disk. Measurements that are appended to it are written to disk instead of being stored in memory. See acquire.expandable_sweepstreamarray() for an example.
Removed extras keyword from core.instantiate().
Added nodes() function to Measurement and MeasurementList that properly-formed instances can use to trace their node path up to the root level.